Extracting a basic screen with some really simple commonalities
